backend/requirements.txt
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0
djangorestframework==3.7.7
httpie==0.9.9
idna==2.6
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.10
MarkupSafe==1.0
mysql-connector-python-rf==2.2.2
mysqlclient==1.3.12
Pygments==2.2.0
pytz==2017.3
requests==2.18.4
uritemplate==3.0.0
urllib3==1.22
